+++ hitl/streaming.py
@@ -6,7 +6,7 @@ import logging
 import ipaddress
 import argparse
 
-from miniconf import Miniconf
+import miniconf
 from stabilizer.stream import measure, StabilizerStream, get_local_ip
 
 logger = logging.getLogger(__name__)
@@ -14,7 +14,7 @@ logger = logging.getLogger(__name__)
 
 async def _main():
     parser = argparse.ArgumentParser(description="Stabilizer Stream HITL test")
-    parser.add_argument("prefix", type=str,
+    parser.add_argument("prefix", type=str, nargs='?',
                         help="The MQTT topic prefix of the target")
     parser.add_argument("--broker", "-b", default="mqtt", type=str,
                         help="The MQTT broker address")
@@ -28,9 +28,19 @@ async def _main():
                         help="Maximum loss for success")
     args = parser.parse_args()
 
+    prefix = args.prefix
+    if not args.prefix:
+        devices = await miniconf.discover(args.broker, 'dt/sinara/dual-iir/+', 1)
+        if not devices:
+            raise Exception('No Stabilizer (Dual-iir) devices found')
+        assert len(devices) == 1, \
+            f'Multiple Stabilizers found: {devices}. Please specify one with --prefix'
+
+        prefix = devices.pop()
+
     logging.basicConfig(level=logging.INFO)
 
-    conf = await Miniconf.create(args.prefix, args.broker)
+    conf = await miniconf.Miniconf.create(prefix, args.broker)
 
     stream_target = [int(x) for x in args.ip.split('.')]
     if ipaddress.ip_address(args.ip).is_unspecified: