+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   not the number of samples per batch. It has been renamed to `batches`.
  * All MQTT clients upgraded and APIs updated.
  * MQTT broker may now be specified via DNS hostnames
+ * `hitl/streaming.py` no longer requires a prefix
+ * Streaming now supports UDP multicast addresses
 
 ## [v0.8.1](https://github.com/quartiq/stabilizer/compare/v0.8.0...v0.8.1) - 2022-11-14)